requirements.txt
ipykernel
# pydantic

# 2.0 Getting Started
#langchain
python-dotenv
ipykernel

# 3.0 Data Ingestion
#langchain-community
ipykernel
#langchain
pypdf
bs4
pymupdf
wikipedia

# 4.0 Data Transformer
langchain-text-splitters

# 5.1 OpenAIEmbedding
#langchain-openai
python-dotenv
# chromadb

# 5.2 OllamaEmbedding
#langchain-community

# 5.3 HuggingFaceEmbedding
sentence-transformers
langchain-huggingface
#langchain
tqdm
# 6.1 Faiss
faiss-cpu
# 6.2 Chroma
# chromadb
# langchain-chroma

# 7.1 Getting Started
#langchain
#langchain-community
#langchain-core
python-dotenv
#langsmith    # Only needed if you're using LangSmith tracking

# 7.2 Simple GenAI App
beautifulsoup4             # HTML parsing
#langchain                 # Core LangChain framework
#langchain-community      # Community integrations (e.g., Ollama, FAISS, loaders)
#langchain-core           # Core building blocks (Document, Prompt, Chains)
faiss-cpu                # FAISS vector database
python-dotenv            # Load environment variables from .env file
ollama                    # Ollama LLM integration
requests                  # HTTP requests

# 8.1 Simple Chatbot App
#langchain
langchain-groq
python-dotenv
# openai
jupyter
langchain_core
#langsmith
requests
tqdm

fastapi       # – for creating API server
uvicorn       # – ASGI server to run FastAPI apps
langserve     # – to convert chains into APIs
python-dotenv # – to load environment variables
starlette     # – (auto dependency of FastAPI, but install if needed)
sse_starlette
# pydantic<2

# 9.1 Chatbot
#langchain
#langchain-core
#langchain-community
langchain-groq
python-dotenv
groq
ipykernel
rich

# 9.1 SimpleLangGraaph
#langchain
langgraph
#langchain-core
#langchain-community
ipykernel

# 9.2 Chatbot
# Core libraries
langgraph
#langchain
#langchain-core
#langchain-community

# OpenAI support
openai
#langchain-openai

# Groq support
langchain-groq

# Ollama support (local LLMs)
ollama

# Env variables
python-dotenv

# 10.2 Pydantic
pydantic

# 10.5 CHATBOT_MULTITOOLS_LANGGRAPH
arxiv
wikipedia

#11-debugging
langgraph-cli[inmem] # lets you build, run, develop, and deploy a LangGraph API server locally






